requirements.txt
# bot training PACKAGES
nltk~=3.8.1
flask~=2.2.3
numpy~=1.23.5
python-dotenv~=1.0.0
tensorflow~=2.12.0;platform_system=="Linux"
tensorflow-macos~=2.12.0;platform_system=="Darwin"
tensorflow-intel~=2.12.0;platform_system=="Windows"
openai==0.27.0
